+++ lib/src/native/audio_management.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import 'media_stream_track_impl.dart';
@@ -70,5 +70,73 @@ class NativeAudioManagement {
       }
     }
     track.enabled = !mute;
+  }
+
+  // ADM APIs
+  static Future<void> startLocalRecording() async {
+    if (!kIsWeb) {
+      try {
+        await WebRTC.invokeMethod(
+          'startLocalRecording',
+          <String, dynamic>{},
+        );
+      } on PlatformException catch (e) {
+        throw 'Unable to start local recording: ${e.message}';
+      }
+    }
+  }
+
+  static Future<void> stopLocalRecording() async {
+    if (!kIsWeb) {
+      try {
+        await WebRTC.invokeMethod(
+          'stopLocalRecording',
+          <String, dynamic>{},
+        );
+      } on PlatformException catch (e) {
+        throw 'Unable to stop local recording: ${e.message}';
+      }
+    }
+  }
+
+  static Future<bool> isVoiceProcessingEnabled() async {
+    if (kIsWeb) return false;
+
+    try {
+      final result = await WebRTC.invokeMethod(
+        'isVoiceProcessingEnabled',
+        <String, dynamic>{},
+      );
+      return result as bool;
+    } on PlatformException catch (e) {
+      throw 'Unable to get isVoiceProcessingEnabled: ${e.message}';
+    }
+  }
+
+  static Future<bool> isVoiceProcessingBypassed() async {
+    if (kIsWeb) return false;
+
+    try {
+      final result = await WebRTC.invokeMethod(
+        'isVoiceProcessingBypassed',
+        <String, dynamic>{},
+      );
+      return result as bool;
+    } on PlatformException catch (e) {
+      throw 'Unable to get isVoiceProcessingBypassed: ${e.message}';
+    }
+  }
+
+  static Future<void> setIsVoiceProcessingBypassed(bool value) async {
+    if (kIsWeb) return;
+
+    try {
+      await WebRTC.invokeMethod(
+        'setIsVoiceProcessingBypassed',
+        <String, dynamic>{"value": value},
+      );
+    } on PlatformException catch (e) {
+      throw 'Unable to set isVoiceProcessingBypassed: ${e.message}';
+    }
   }
 }