+++ lib/src/native/media_stream_impl.dart
@@ -2,7 +2,7 @@
 import 'dart:async';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import 'factory_impl.dart';