+++ lib/src/native/mediadevices_impl.dart
@@ -5,7 +5,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 
 // Package imports:
-import 'package:webrtc_interface_plus/webrtc_interface_plus.dart';
+import 'package:webrtc_interface/webrtc_interface.dart';
 
 // Project imports:
 import 'event_channel.dart';